feat(agent): Continue work on agent refactor [wip]
Finishes implementation of new agent models and proxies, static agent registry, and instance agents.
Still need to implement buildRevertFunction and GameInstance.using
