--- src/new-agents/agent-manager.ts
@@ -1,5 +1,0 @@
-export class AgentManager {}
-
-export const isAgentManager = (o: any): o is AgentManager => {
-    return true;
-};
