--- src/new-agents/todo.ts
@@ -1,20 +1,0 @@
-import { Agent } from "./agent-model";
-
-export const isContextStatic = () => {
-    return true;
-};
-
-export const getNextStaticAgentId = () => {
-    return 1;
-};
-
-export const addAgentToStaticRegistry = (agent: Agent) => {
-    return;
-};
-
-export const getStaticAgentProperty = (
-    id: number,
-    property: PropertyKey
-): any => {
-    return;
-};
